+++ core/src/androidMain/kotlin/BuildConfig.kt
@@ -1,5 +1,5 @@
 @Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
-public actual object BuildConfig {
-    public actual val isDebug: Boolean
+internal actual object BuildConfig {
+    actual val isDebug: Boolean
         get() = com.powersync.BuildConfig.DEBUG
 }