+++ core/src/commonIntegrationTest/kotlin/com/powersync/testutils/TestUtils.kt
@@ -100,7 +100,6 @@ internal class ActiveDatabaseTest(
             everySuspend { getCredentialsCached() } returns
                 PowerSyncCredentials(
                     token = "test-token",
-                    userId = "test-user",
                     endpoint = "https://test.com",
                 )