+++ core/src/commonMain/kotlin/BuildConfig.kt
@@ -1,4 +1,4 @@
 @Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
-public expect object BuildConfig {
-    public val isDebug: Boolean
+internal expect object BuildConfig {
+    val isDebug: Boolean
 }