+++ core/src/commonMain/kotlin/com/powersync/db/Queries.kt
@@ -5,6 +5,8 @@ import com.powersync.db.internal.ConnectionContext
 import com.powersync.db.internal.PowerSyncTransaction
 import kotlinx.coroutines.flow.Flow
 import kotlin.coroutines.cancellation.CancellationException
+import kotlin.time.Duration
+import kotlin.time.Duration.Companion.milliseconds
 
 public fun interface ThrowableTransactionCallback<R> {
     @Throws(PowerSyncException::class, kotlinx.coroutines.CancellationException::class)
@@ -17,8 +19,21 @@ public fun interface ThrowableLockCallback<R> {
 }
 
 public interface Queries {
+    public companion object {
+        /**
+         * The default throttle duration for  [onChange] and [watch] operations.
+         */
+        public val DEFAULT_THROTTLE: Duration = 30.milliseconds
+    }
+
     /**
-     * Execute a write query (INSERT, UPDATE, DELETE)
+     * Executes a write query (INSERT, UPDATE, DELETE).
+     *
+     * @param sql The SQL query to execute.
+     * @param parameters The parameters for the query, or an empty list if none.
+     * @return The number of rows affected by the query.
+     * @throws PowerSyncException If a database error occurs.
+     * @throws CancellationException If the operation is cancelled.
      */
     @Throws(PowerSyncException::class, CancellationException::class)
     public suspend fun execute(
@@ -27,9 +42,14 @@ public interface Queries {
     ): Long
 
     /**
-     * Execute a read-only (SELECT) query and return a single result.
-     * If there is no result, throws an [IllegalArgumentException].
-     * See [getOptional] for queries where the result might be empty.
+     * Executes a read-only (SELECT) query and returns a single result.
+     *
+     * @param sql The SQL query to execute.
+     * @param parameters The parameters for the query, or an empty list if none.
+     * @param mapper A function to map the result set to the desired type.
+     * @return The single result of the query.
+     * @throws PowerSyncException If a database error occurs or no result is found.
+     * @throws CancellationException If the operation is cancelled.
      */
     @Throws(PowerSyncException::class, CancellationException::class)
     public suspend fun <RowType : Any> get(
@@ -39,7 +59,14 @@ public interface Queries {
     ): RowType
 
     /**
-     * Execute a read-only (SELECT) query and return the results.
+     * Executes a read-only (SELECT) query and returns all results.
+     *
+     * @param sql The SQL query to execute.
+     * @param parameters The parameters for the query, or an empty list if none.
+     * @param mapper A function to map the result set to the desired type.
+     * @return A list of results.
+     * @throws PowerSyncException If a database error occurs.
+     * @throws CancellationException If the operation is cancelled.
      */
     @Throws(PowerSyncException::class, CancellationException::class)
     public suspend fun <RowType : Any> getAll(
@@ -49,7 +76,14 @@ public interface Queries {
     ): List<RowType>
 
     /**
-     * Execute a read-only (SELECT) query and return a single optional result.
+     * Executes a read-only (SELECT) query and returns a single optional result.
+     *
+     * @param sql The SQL query to execute.
+     * @param parameters The parameters for the query, or an empty list if none.
+     * @param mapper A function to map the result set to the desired type.
+     * @return The single result of the query, or null if no result is found.
+     * @throws PowerSyncException If a database error occurs.
+     * @throws CancellationException If the operation is cancelled.
      */
     @Throws(PowerSyncException::class, CancellationException::class)
     public suspend fun <RowType : Any> getOptional(
@@ -59,65 +93,93 @@ public interface Queries {
     ): RowType?
 
     /**
-     * Execute a read-only (SELECT) query every time the source tables are modified and return the results as a List in [Flow].
+     * Returns a [Flow] that emits whenever the source tables are modified.
+     *
+     * @param tables The set of tables to monitor for changes.
+     * @param throttleMs The minimum interval, in milliseconds, between emissions. Defaults to [DEFAULT_THROTTLE]. Table changes are accumulated while throttling is active. The accumulated set of tables will be emitted on the trailing edge of the throttle.
+     * @param triggerImmediately If true (default), the flow will immediately emit an empty set of tables when the flow is first collected. This can be useful for ensuring that the flow emits at least once, even if no changes occur to the monitored tables.
+     * @return A [Flow] emitting the set of modified tables.
+     * @throws PowerSyncException If a database error occurs.
+     * @throws CancellationException If the operation is cancelled.
+     */
+    @Throws(PowerSyncException::class, CancellationException::class)
+    public fun onChange(
+        tables: Set<String>,
+        throttleMs: Long = DEFAULT_THROTTLE.inWholeMilliseconds,
+        triggerImmediately: Boolean = true,
+    ): Flow<Set<String>>
+
+    /**
+     * Executes a read-only (SELECT) query every time the source tables are modified and returns the results as a [Flow] of lists.
+     *
+     * @param sql The SQL query to execute.
+     * @param parameters The parameters for the query, or an empty list if none.
+     * @param throttleMs The minimum interval, in milliseconds, between queries. Defaults to [DEFAULT_THROTTLE].
+     * @param mapper A function to map the result set to the desired type.
+     * @return A [Flow] emitting lists of results.
+     * @throws PowerSyncException If a database error occurs.
+     * @throws CancellationException If the operation is cancelled.
      */
     @Throws(PowerSyncException::class, CancellationException::class)
     public fun <RowType : Any> watch(
         sql: String,
         parameters: List<Any?>? = listOf(),
-        /**
-         * Specify the minimum interval, in milliseconds, between queries.
-         */
-        throttleMs: Long? = null,
+        throttleMs: Long = DEFAULT_THROTTLE.inWholeMilliseconds,
         mapper: (SqlCursor) -> RowType,
     ): Flow<List<RowType>>
 
     /**
-     * Takes a global lock, without starting a transaction.
+     * Takes a global lock without starting a transaction.
      *
-     * This takes a global lock - only one write transaction can execute against
-     * the database at a time. This applies even when constructing separate
-     * database instances for the same database file.
-     *
-     * Locks for separate database instances on the same database file
-     * may be held concurrently.
+     * This lock ensures that only one write transaction can execute against the database at a time, even across separate database instances for the same file.
      *
      * In most cases, [writeTransaction] should be used instead.
+     *
+     * @param callback The callback to execute while holding the lock.
+     * @return The result of the callback.
+     * @throws PowerSyncException If a database error occurs.
+     * @throws CancellationException If the operation is cancelled.
      */
     @Throws(PowerSyncException::class, CancellationException::class)
     public suspend fun <R> writeLock(callback: ThrowableLockCallback<R>): R
 
     /**
-     * Open a read-write transaction.
+     * Opens a read-write transaction.
      *
-     * This takes a global lock - only one write transaction can execute against
-     * the database at a time. This applies even when constructing separate
-     * database instances for the same database file.
+     * This takes a global lock, ensuring that only one write transaction can execute against the database at a time, even across separate database instances for the same file.
      *
-     * Statements within the transaction must be done on the provided
-     * [PowerSyncTransaction] - attempting statements on the database
-     * instance will error cause a dead-lock.
+     * Statements within the transaction must be done on the provided [PowerSyncTransaction] - attempting statements on the database instance will error cause a dead-lock.
+     *
+     * @param callback The callback to execute within the transaction.
+     * @return The result of the callback.
+     * @throws PowerSyncException If a database error occurs.
+     * @throws CancellationException If the operation is cancelled.
      */
     @Throws(PowerSyncException::class, CancellationException::class)
     public suspend fun <R> writeTransaction(callback: ThrowableTransactionCallback<R>): R
 
     /**
-     * Takes a read lock, without starting a transaction.
+     * Takes a read lock without starting a transaction.
      *
-     * The lock only applies to a single SQLite connection, and multiple
-     * connections may hold read locks at the same time.
+     * The lock applies only to a single SQLite connection, allowing multiple connections to hold read locks simultaneously.
      *
-     * In most cases, [readTransaction] should be used instead.
+     * @param callback The callback to execute while holding the lock.
+     * @return The result of the callback.
+     * @throws PowerSyncException If a database error occurs.
+     * @throws CancellationException If the operation is cancelled.
      */
     @Throws(PowerSyncException::class, CancellationException::class)
     public suspend fun <R> readLock(callback: ThrowableLockCallback<R>): R
 
     /**
-     * Open a read-only transaction.
+     * Opens a read-only transaction.
      *
-     * Statements within the transaction must be done on the provided
-     * [PowerSyncTransaction] - executing statements on the database level
-     * will be executed on separate connections.
+     * Statements within the transaction must be done on the provided [PowerSyncTransaction] - executing statements on the database level will be executed on separate connections.
+     *
+     * @param callback The callback to execute within the transaction.
+     * @return The result of the callback.
+     * @throws PowerSyncException If a database error occurs.
+     * @throws CancellationException If the operation is cancelled.
      */
     @Throws(PowerSyncException::class, CancellationException::class)
     public suspend fun <R> readTransaction(callback: ThrowableTransactionCallback<R>): R