+++ core/src/commonMain/kotlin/com/powersync/sync/SyncStream.kt
@@ -179,7 +179,6 @@ internal class SyncStream(
                 contentType(ContentType.Application.Json)
                 headers {
                     append(HttpHeaders.Authorization, "Token ${credentials.token}")
-                    append("User-Id", credentials.userId ?: "")
                 }
             }
         if (response.status.value == 401) {
@@ -207,7 +206,6 @@ internal class SyncStream(
                     contentType(ContentType.Application.Json)
                     headers {
                         append(HttpHeaders.Authorization, "Token ${credentials.token}")
-                        append("User-Id", credentials.userId ?: "")
                     }
                     timeout { socketTimeoutMillis = Long.MAX_VALUE }
                     setBody(bodyJson)