+++ core/src/commonTest/kotlin/com/powersync/sync/SyncStreamTest.kt
@@ -77,7 +77,6 @@ class SyncStreamTest {
                 everySuspend { getCredentialsCached() } returns
                     PowerSyncCredentials(
                         token = "test-token",
-                        userId = "test-user",
                         endpoint = "https://test.com",
                     )
             }