+++ CHANGELOG.md
@@ -1,9 +1,30 @@
 # Changelog
 
-## unreleased
+## 1.1.0
 
+* Add `trackPreviousValues` option on `Table` which sets `CrudEntry.previousValues` to previous values on updates.
+* Add `trackMetadata` option on `Table` which adds a `_metadata` column that can be used for updates.
+  The configured metadata is available through `CrudEntry.metadata`.
+* Add `ignoreEmptyUpdates` option which skips creating CRUD entries for updates that don't change any values.
+
+## 1.0.1
+
+* [Internal] Version bump for broken Swift release pipeline
+
+## 1.0.0
+
+* Bump SDK to V1/Stable feature status
+* Fixed `CrudBatch` `hasMore` always returning false.
+* Added `triggerImmediately` to `onChange` method.
 * Report real-time progress information about downloads through `SyncStatus.downloadProgress`.
 * Compose: Add `composeState()` extension method on `SyncStatus`.
+* [Internal] Added helper method for Swift `PowerSyncException` throwing.
+
+## 1.0.0-BETA32
+
+* Added `onChange` method to the PowerSync client. This allows for observing table changes.
+* Removed unnecessary `User-Id` header from internal PowerSync service requests.
+* Fix loading native PowerSync extension for Java targets.
 
 ## 1.0.0-BETA31